+++ src/Trimmer/wrapper.jl
@@ -5,7 +5,7 @@ function julia_wrapper_atria(ARGS::Vector{String}; exit_after_help = true)
 
     time_program_initializing = time()
 
-    atria_version = "v3.0.0"
+    atria_version = "v3.0.1"
 
     args = parsing_args(ARGS; ver = atria_version, exit_after_help = exit_after_help)
 
@@ -579,20 +579,6 @@ function julia_wrapper_atria(ARGS::Vector{String}; exit_after_help = true)
         empty!(r1s)
         empty!(r2s)
 
-        #=
-        global n_reads
-        global n_r1
-        global n_r2
-        global nbatch
-        global total_read_copied_in_loading
-        global total_n_goods
-        global total_n_reads
-        global total_n_bytes_read1
-        global total_n_bytes_read2
-        global chunk_size1
-        global chunk_size2
-        =#
-
         n_reads = 0
         n_r1 = 0
         n_r2 = 0
@@ -646,9 +632,7 @@ function julia_wrapper_atria(ARGS::Vector{String}; exit_after_help = true)
             processing_reads_threads!(r1s, r2s, isgoods, n_reads)
 
             isgoods_in_range = view(isgoods, 1:n_reads)
-            n_goods = sum(isgoods_in_range)
-            total_n_goods += n_goods
-            total_n_reads += n_reads
+            task_sum = Threads.@spawn sum(isgoods_in_range)
 
             write_fqs_threads!(
                 io1out::IO, io2out::IO,
@@ -656,7 +640,13 @@ function julia_wrapper_atria(ARGS::Vector{String}; exit_after_help = true)
                 r1s::Vector{FqRecord}, r2s::Vector{FqRecord},
                 n_reads::Int, isgoods_in_range)
 
-            @info "Cycle $nbatch: processed $n_reads read pairs ($total_n_reads in total), in which $n_goods passed filtration ($total_n_goods in total). ($ncopied/$total_read_copied_in_loading reads copied)"
+            n_goods = fetch(task_sum)
+            total_n_goods += n_goods
+            total_n_reads += n_reads
+
+            # @info "Cycle $nbatch: processed $n_reads read pairs ($total_n_reads in total), in which $n_goods passed filtration ($total_n_goods in total). ($ncopied/$total_read_copied_in_loading reads copied)"
+
+            @info "Cycle $nbatch: read $n_reads/$total_n_reads pairs; wrote $n_goods/$total_n_goods pairs; (copied $ncopied/$total_read_copied_in_loading reads)"
         end
 
         while !eof(io1::IO) || !eof(io2::IO)